+++ SRTShareLib/SRTManager/ProtocolManager.cs
@@ -6,7 +6,7 @@ namespace SRTShareLib
 {
     public class ProtocolManager
     {
-        public const byte DEFAULT_QUALITY = 50;  // video default quality value
+        public const long DEFAULT_QUALITY = 50L;  // video default quality value
 
         /// <summary>
         /// The function generates a syn cookie for the handshake
